master Race finished

+++ backend/database/migrations/2019_09_01_101054_create_races_table.php
@@ -15,7 +15,7 @@ class CreateRacesTable extends Migration
 	{
 		Schema::create('races', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->integer('best_time')->nullable();
+			$table->float('best_time')->nullable();
 			$table->string('status')->nullable();
 			$table->integer('current_time')->nullable();
 			$table->integer('completed_horse_count')->nullable();
